requirements.txt
tensorflow
tensorflow_probability
scikit-learn
xgboost
torch
torchvision
torchaudio
keras
tf-keras
pandas
numpy
scipy
statsmodels
matplotlib
seaborn
xbbg
python-dotenv
pickle